Remove sign and move relevant code to wallet
