+++ src/wallet/index.ts
@@ -1,18 +1,25 @@
 import { fromSeed } from 'bip32'
 import { mnemonicToSeedSync } from 'bip39'
+import { assert } from 'chai'
 import { classicAddressToXAddress } from 'ripple-address-codec'
-import { decode, encodeForSigning } from 'ripple-binary-codec'
+import {
+  decode,
+  encodeForSigning,
+  encodeForMultisigning,
+  encode,
+} from 'ripple-binary-codec'
 import {
   deriveAddress,
   deriveKeypair,
   generateSeed,
   verify,
+  sign,
 } from 'ripple-keypairs'
+import { computeSignedTransactionHash } from '..'
 
 import ECDSA from '../common/ecdsa'
 import { ValidationError } from '../common/errors'
 import { Transaction } from '../models/transactions'
-import { signOffline } from '../transaction/sign'
 import { SignOptions } from '../transaction/types'
 
 const DEFAULT_ALGORITHM: ECDSA = ECDSA.ed25519
@@ -188,6 +195,121 @@ class Wallet {
   public getClassicAddress(): string {
     return deriveAddress(this.publicKey)
   }
+}
+
+function signWithKeypair(
+  txJSON: string,
+  publicKey: string,
+  privateKey: string,
+  options: SignOptions = {
+    signAs: '',
+  },
+): { signedTransaction: string; id: string } {
+  const tx = JSON.parse(txJSON)
+  if (tx.TxnSignature || tx.Signers) {
+    throw new ValidationError(
+      'txJSON must not contain "TxnSignature" or "Signers" properties',
+    )
+  }
+
+  const txToSignAndEncode = { ...tx }
+
+  txToSignAndEncode.SigningPubKey = options.signAs ? '' : publicKey
+
+  if (options.signAs) {
+    const signer = {
+      Account: options.signAs,
+      SigningPubKey: publicKey,
+      TxnSignature: computeSignature(
+        txToSignAndEncode,
+        privateKey,
+        options.signAs,
+      ),
+    }
+    txToSignAndEncode.Signers = [{ Signer: signer }]
+  } else {
+    txToSignAndEncode.TxnSignature = computeSignature(
+      txToSignAndEncode,
+      privateKey,
+    )
+  }
+  const serialized = encode(txToSignAndEncode)
+  checkTxSerialization(serialized, tx)
+  return {
+    signedTransaction: serialized,
+    id: computeSignedTransactionHash(serialized),
+  }
+}
+
+/**
+ *  Decode a serialized transaction, remove the fields that are added during the signing process,
+ *  and verify that it matches the transaction prior to signing.
+ *
+ * @param serialized - A signed and serialized transaction.
+ * @param tx - The transaction prior to signing.
+ *
+ * @returns This method does not return a value, but throws an error if the check fails.
+ */
+function checkTxSerialization(serialized: string, tx: Transaction): void {
+  // Decode the serialized transaction:
+  const decoded = decode(serialized)
+
+  // ...And ensure it is equal to the original tx, except:
+  // - It must have a TxnSignature or Signers (multisign).
+  if (!decoded.TxnSignature && !decoded.Signers) {
+    throw new ValidationError(
+      'Serialized transaction must have a TxnSignature or Signers property',
+    )
+  }
+  // - We know that the original tx did not have TxnSignature, so we should delete it:
+  delete decoded.TxnSignature
+  // - We know that the original tx did not have Signers, so if it exists, we should delete it:
+  delete decoded.Signers
+
+  // - If SigningPubKey was not in the original tx, then we should delete it.
+  //   But if it was in the original tx, then we should ensure that it has not been changed.
+  if (!tx.SigningPubKey) {
+    delete decoded.SigningPubKey
+  }
+
+  // - Memos have exclusively hex data which should ignore case.
+  //   Since decode goes to upper case, we set all tx memos to be uppercase for the comparison.
+  tx.Memos?.map((memo) => {
+    if (memo.Memo.MemoData) {
+      memo.Memo.MemoData = memo.Memo.MemoData.toUpperCase()
+    }
+
+    if (memo.Memo.MemoType) {
+      memo.Memo.MemoType = memo.Memo.MemoType.toUpperCase()
+    }
+
+    if (memo.Memo.MemoFormat) {
+      memo.Memo.MemoFormat = memo.Memo.MemoFormat.toUpperCase()
+    }
+
+    return memo
+  })
+
+  assert.deepEqual(
+    decoded as unknown as Transaction,
+    tx,
+    'Serialized transaction does not match the original transaction.',
+  )
+}
+
+function computeSignature(tx: object, privateKey: string, signAs?: string) {
+  const signingData = signAs
+    ? encodeForMultisigning(tx, signAs)
+    : encodeForSigning(tx)
+  return sign(signingData, privateKey)
+}
+
+function signOffline(
+  wallet: Wallet,
+  txJSON: string,
+  options?: SignOptions,
+): { signedTransaction: string; id: string } {
+  return signWithKeypair(txJSON, wallet.publicKey, wallet.privateKey, options)
 }
 
 export default Wallet
